guardando o numero de créditos de eletivas na cache

+++ app/src/main/java/com/devmob/minhagrade/Model/Prefs.java
@@ -28,16 +28,16 @@ public class Prefs {
     }
 
     // Metodo para salvar int
-    public static void setInteger(Context context, String chave, int valor){
+    public static void setFloat(Context context, String chave, Float valor){
         SharedPreferences pref = context.getSharedPreferences(MY_PREFS_NAME,context.MODE_PRIVATE);
         SharedPreferences.Editor editor = pref.edit();
-        editor.putInt(chave,valor);
+        editor.putFloat(chave,valor);
         editor.apply();
     }
     // Metodo para buscar int
-    public static int getInt(Context context, String chave){
+    public static Float getFloat(Context context, String chave){
         SharedPreferences pref = context.getSharedPreferences(MY_PREFS_NAME,context.MODE_PRIVATE);
-        int valor = pref.getInt(chave,0);
+        Float valor = pref.getFloat(chave,0);
         return valor;
     }
